Fleshed out move and scan functions
The simulator now uses coordinate geometry on a Cartesian plane to
determine precisely where on the map it is being asked to go, if it can
reach that destination, and can move to that location or as close as
possible.

It can also initiate a scan and determine the distance to the nearest
object at any given angle, though some inaccuracies remain, specifically
regarding the limitations of the actual laser.

Code for the rotation functionality is present, but due to time
limitations has not been robustly tested. Stopping remains
non-functional. I will re-factor the code at some point after this
sprint. Finally, the simulator still cannot communicate with any other
code at present.