Re-factored code
Most of the computational code has been moved away from the main .ino
file to the included libraries. The serial communication code, however,
remains, due to an odd bug when run from a library. (It will likely be
removed entirely at a future date, anyways, to be replaced with an SPI
implementation.)