Made command handlers fast enough for interrupts
Relocated command processing code out of the command handlers to its own
functions, as the command handlers are called by the SPI_Wrapper library
as part of an interrupt, and must therefore execute quickly. The command
handlers now store the received command in an object on the heap, and
the command is processed on the next loop(). (Currently, the object on
the heap is not being deleted when no longer in use. This will be
remedied at a later stage.)