Implemented real-time emulation
The simulator should now emulate how long the real robot would take to
move to a given position. It does this by creating a timer and waiting
until the time has elapsed before sending the response to the Edison.

I have also written a function to calculate the distance that has been
travelled, based on the time elapsed and the robot's speed. This will be
used when the stop functionality has been successfully implemented, to
determine the robot's position upon halting.

Finally, I have updated the values of various distance variables, in
light of the requirement that distances be measured in centimetres (I
had been treating them as metres).