Provided for deletion of objects not in use
The simulator now deletes the command object from the heap once it is no
longer needed.

I have also corrected the representation of the angle (the documentation
is outdated; angle is measured in degrees, not deci-degrees) and moved
most of the scan processing code to the loop, where it seems more
appropriate.

Finally, I have commented out a significant amount of older code, which
will be removed entirely once I am certain it will no longer be needed
for reference.