Corrected object deletion & command processing in loop
The command object on the heap is now properly deleted, and the global
command pointer variable is now set back to NULL at the same time.

Also, the if statement that determines when to process a received
command now checks that the robot is not moving or rotating (otherwise
the simulator would continuously attempt to process the command and
never send a response over SPI).