Possibly fix bugs described in issue
The main fault was in the hasInterception function. If neither the wall
in question nor the line generated by the scanner's "beam" were
vertical, the function would check that the "intercept" variable was
within the range 0.0 >= intercept <= 300.0 (i.e. within the bounds of
the room) and return accordingly. However, this intercept variable was
calculated as a y-intercept based on the assumption that the x value was
known, which only holds true for vertical lines. This caused scanning
lines between the degrees 50 -> 90, 95 -> 130, 230 -> 265, and 275 ->
315 to return false, because these values should have intersected the
"East" and "West" walls of the virtual room, which are horizontal lines.
I have removed the boundary check.

Subsequent testing revealed problems when scanning at the angles 45, 90,
135, 180, 225, 241, 270, 315 and 360. The majority of these were caused
by a common difficulty with floating-point numbers: certain calculations
that should have given a zero result instead gave an infinitesimally
small, non-zero result. I have added tolerance-based comparison tests to
account for this. One was caused by a silent index-out-of-bounds
condition, which I have rectified.

Finally, I have reduced the default point's distance from 65530.0 to
4096.0. While this would still cause issues if another bug caused no
other intersection points to be found, the new value should avoid any
potential problems with value overflow (i.e. 65535).

(Note: there is a substantial amount of debugging code present. I am
retaining it in this commit, so that it can be referenced in the event
that the bugs have not been fixed, but I will remove it in the next
commit, in anticipation of the bugs being resolved.)