Create room class
The creation of this room class should allow for more flexible and
complex room topologies. In particular, the room can now contain
objects, as opposed to being restricted to an empty room.

The structures that were previously in structs.h have been transposed to
the room class; MapLine and EquationOfLine have been merged into the one
Line class.
An Object consists of an integer for the number of sides the object has,
and a pointer to an array of Lines of the appropriate size.
A room consists of a Point representing the co-ordinates of the robot,
an Object
representing the walls of the room, an integer tracking the number of
objects the room can contain, and a pointer to an array of Objects of
the appropriate size. There is also a function to add an object to the
room.

Finally, the simulator and calc library have been modified to work with
the room class.

References #146.