Add timer-based delay to first scanResponse
This appears to be why the simulator appeared to fail to respond to a
scanCommand from the Edison. We had thought that the simulator may have
been overwhelming the Edison with too many scanResponses, but the timer
interval seemed sufficiently large.
However, it was not being applied to the first scanResponse, which was
being sent immediately on receipt of the scanCommand - apparently before
the Edison was ready.
If it misses that first scanResponse, the Edison does not receive any
subsequent scanResponses. I believe it was a fluke that the Edison ever
received a set of scanResponses.

The possibility remains that the simulator does send scanResponses too
quickly, causing the Edison to miss a scanResponse and hang; however, an
interval of 80ms seemed to work during testing, so the interval was
increased to 100ms for conservatism's sake. It can be increased further,
if necessary.

References #161.