Add compass to simulator
Note: Because a heading of 0 degrees in the simulator is actually due
West, the compass heading is returned as the physicalAngle value less 90
modulo 360.

The physical angle is also subject to error, when enabled (since, on the
physical robot, it would be the wheels slipping and rotating the robot
that would cause a move to go awry).

uniqueID has been made actually unique (at least per run-time).

References #193.