Adjust value of angleSlip
Due to the nature of rounding and integers, at the old value of
angleSlip, it would require a degree of error above 33 percent to cause
the robot to stray 1 degree off course. However, that degree of error
would mean that the robot would only travel 66 percent of the distance,
which is disproportionate.
At the new value, it will only require a degree of error of 5 percent to
cause the same amount of drift, which is more reasonable.

Also, I don't know why that cast to float was there, but it's redundant.