Fix edge case allowing illegal destinations
Consider the following scenario:

1. The simulator executes a move command, but encounters a wall or
object, stops moving and sets its current position accordingly. The
robot is now on the line of the wall or side of object.
2. Another move command is issued, in either the same direction or
another which would direct the robot through the wall or side of object.

The wall or side of object would be detected, and a valid interception
point would be generated, but it would be identical to the robot's
current position, and because therefore there would be no translation in
x or y when moving to that point, it would not be set as the nearest
point to the robot. If no other point could be found that would be
closer than the projected destination, that original destination would
be returned as the correct destination, even if it would take the robot
out of the map or through an object.

The code has been modified to compare each valid interception point
against the robot's current position, and if a match is found, the
simulator "steps back" from the wall or side of object, in order to
obtain a translation in x and y towards it, and determines if the signs
of translation towards the wall or side of object and towards the
projected destination are identical. If so, it returns the robot's
current position as the destination (i.e. no movement).

References #162.