Account for mirroring of directions in simulator
In the unit circle, angles are measured from the X-axis,
counter-clockwise. This means that the angles 90 and 270 degrees (and
similar angles in each direction) are in the inverse positions from the
compass convention.

(To illustrate: draw and label the unit circle, then rotate the diagram
so that 0 degrees faces "up", as on a compass. 90 degrees (East) will
face "left", or conventional West, and vice versa.)