Add ability to easily switch between rooms
Previously, changing the parameters of the room required carefully
modifying a number of variables. This was awkward, so it was requested
that it be made possible to load different pre-defined rooms based on
one variable.
Initially, the code was contained in the setup function in the main
sketch, but it became immediately apparent that it would look messy and
make the main code less maintainable if left there, so a new class was
created to hold the new code.
Switching between rooms is now accomplished by passing one variable to a
function in the new class.

To create a Room object without explicitly initializing it, a default
constructor was required, so one was written.

The currentPosition variable is no longer needed, since a Room object
has the robot's current position as a member variable.

References #163, #164.