Reduce occurrence of very large scan distances
The room class now has a farthest point, which takes as its parameters
the largest possible value for x and y for that room, even if that point
would not be a legal location in that room (for example, a triangular
room with a base of 300 cm and a height of 300 cm would have a farthest
point of (300, 300), despite that location being invalid).
Then, in the main sketch, there is a new unsigned long, MAX_DISTANCE,
which is the distance from the origin to the room's farthest point. If a
distance returned by a scan is greater than MAX_DISTANCE, the data is
discarded.